backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Data Validation & Settings
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Database & ORM
sqlalchemy==2.0.36
sqlmodel==0.0.22
alembic==1.13.2
asyncpg==0.30.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.12

# HTTP Client
httpx==0.27.2

# Payment Processing
stripe==10.8.0

# Supabase
supabase==2.8.0

# Utilities
tenacity==9.0.0
sentry-sdk==2.18.0
geoip2==4.8.0
user-agents==2.2.0
email-validator==2.2.0

# Storage
b2sdk==2.1.0

# Task Scheduling
apscheduler==3.10.4